requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
Pillow==8.3.2
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
gunicorn==20.1.0
whitenoise==6.2.0
mysqlclient==2.1.1
djangorestframework
django-cors-headers